requirements.txt
flask==3.0.0
openai==1.3.0
fastmcp==0.1.0
python-dotenv==1.0.0
asyncio==3.4.3